pyproject.toml
[tool.poetry]
name = "py-learning"
version = "0.1.0"
description = "this is description"
authors = ["快乐的小梁同学 <350442340@qq.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlmodel = "^0.0.22"
uvicorn = "^0.32.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
